Add model Address to app people and include interface in admin of ArtaUser
Create the model in address.py and register it with __init__.py.
We need to migrate to get the database table.
We make an inline admin interface that we use in the newly created ArtaUserAdmin class so that an address is displayed in the same admin page as the rest of the user fields.