views: Add three placeholder views and their templates
Add a basic view for the general entry page (for now temporarily added in the
views.py of the people app), one for the entry page for the people app
and one for the entry page of the event app.
Register these views via the respective urls.py and register the app specific
urls with the arta/urls.py.
Also add corresponding templates. (contain placeholder text)