django-allauth: Use for authentication
This enables the Google backend for now, but additional steps are needed
to set that up and store the needed keys, see:

	https://django-allauth.readthedocs.io/en/latest/providers.html#google

Allauth needs the django.contrib.sites app, but it's not entirely clear
how this is used exactly and if it is really needed for running just a
single domain.

Allauth defines some urls under /account to manage your email addresses,
password, external authentications, etc. These should probably be better
integrated later.