registrations: limit admin dropdowns for registration options
RegistrationField and RegistrationFieldOption can depend on other
RegistrationFieldOptions. By default, all of them are shown in the
admin, but only the ones related to the same event are relevant.

There seems to be no builtin way to limit this, so this overrides some
methods on the admin class to achieve the same. Not entirely pretty, but
it works.