reversion: Follow relations around registrations and events
This lets reversion include all RegistrationFields and
RegistrationFieldOptions along with their Event, and
RegistrationFieldValues along with their Registration, so reverting to
an older revision of an Event or Registration includes the fields as
you would expect.

The reverse also happens: when a field/option/value is changed, the
revision for that will include the Event or Registration (and thus also
all related fields/options/values) so the change shows up in the history
for the Event or Registration.