Switch from pipenv to poetry
This is an alternative dependency manager that has similar capabilities,
but seems to have a better development process. A particular problem
with pipenv is that somewhat enforces updating all dependencies whenever
installing something (partly intentional, partly due to unfixed bugs).
Poetry only updates when asked and can additionaly also manage and
install library projects (not relevant here, but useful to not need
another tool for libraries).

This also updates and expands the README a bit wrt to setup and
updating.