RegistrationField: Include event in natural key
Previously, only the field name was used, but the field name is not
globally unique, only per event.

This als changes the natural key of RegistrationFieldOption, which now
includes the natural key of RegistrationField as a sublist
(e.g. [[On the Styx, type], Player] instead of [type, Player]). Using a
means that further changes to the natural key for RegistrationField will
be automatically reflected in RegistrationFieldOption.