Site structure: change home view into dashboard views
Add a dashboard view to core that shows a short welcome message and (WIP) lists
of events to register for and events in the future that have been registered
for (these lists are not filled yet: WIP).
Remove the former home.html.
Remove the old home view that lived in the people app. Move it (the new one) to
the core app.
Include the right urls in core/urls.py and remove the explicit include of the
home page view from people_views. (this was taken care of in an earlier commit,
including the core urls).
Move (and rename) the link to the home/dashboard page to within the 'is
authenticated' check in base.html.