Rename css classes and turn span into div
We need block level elements for better styling so turn span into div.
Also rename seriesname and eventname to the more general bigname en subname
which is more in accordance with their (future) styling. Also make sure that if
there is not series name, that the event name becomes the 'big name'.