registrations: Show e-mailaddress during registration
Changing it during the registration flow is complex (because it must be
verified before making it primary), so instead the e-mail address is
only shown, with a reference to the account settings to change it.

By default, a disabled field shows a disabled input box, which is not so
nice (and might truncate the e-mailaddress), so this adds a small widget
that just shows the e-mailaddress in a span instead.

This relates to #81.