Remove link to personal detail page
For now remove the link to the page with personal details (name/address,
medical info) for we have no way for the user to modify this yet.
And when we do implement an option for modification we have to add a 'keep this
information' feature that may need some more thoughts in #79. Also might touch
#42 (medical details)