template: show event url when suggestion to go there
In the registration_confirmation page we suggest the user to check info on the
event site, so now we also provide them with a link.
Unfortunately it is hard to turn text neatly into a link inside a blocktrans
tag, so we just put the whole url (but clickable by urlize) in the text.
If there is no url to display nothing will be added to the text.