Registration wizard: change order of steps
Move the events options step to be the first so people can decide if they like
the options and the price(s) before adding personal details and later deciding
that they do not want to register after all.

(This also fixes a bug of a continues wizard loop when an event has no options.
It did not set preregistration_completed status when there were no options,
redirected to final check that redirected back to the first step (personal
details) in the old order of things)