settings: Prevent breakage without mysqlclient installed
In commit df1964a (Backport Django commit to allow CheckConstraint on
MariaDB), some code was added to the common settings that imports the
mysql backend. However, this will in turn try to import the mysqlclient
package and break if not installed.

This commit simply catches the resulting exception and ignores it.