Event listings: add registration info block
Besides an event detail collapsable block we now also have a collapsable block
with information about the registration (of course only for events user has
registered for).
To keep the layout comprehensible we move the 'event details' "tab" to the left
and add the 'registration details' "tab" on the right when needed.
We need some additional classes and ids in the html for this and some
additional styling to get the event details tab to look good.
Also made sure that only one info div (for each event) can be open at a time.