MedicalDetailForm: Ensure that MedicalDetails and ConsentLog are modified atomically
This prevents a situation where the ConsentLog entry could not be
created, but the MedicalDetails are created (but without tracking the
consent for that).

In practice, this was already prevented because the view runs inside a
revision, which is atomic, but putting another atomic block in the form
makes the guarantee strong.

This also adds a test for this.