RegistrationStatusService: Optimize Event used_slots query
Previously, we used one query to lock the event and then another to
retrieve the event with its used slots, selecting the event data twice.

Now, the first query locks the event and selects its data and the second
query only retrieves the number of used slots. This should make the
second query slightly faster and prevents having to construct a second
unused Event object.

Note that using a separate query for locking was done to support
Postgresql, but it seems this is also useful to make sure only the event
is locked, rather than all rows used in the `with_slots()` query (i.e.
*all* registrations for that event). This commit also updates the
comment to reflect this.