RegistrationNotifyService: Use request.user when possible
Previously, this would always use registration.user, which would be
loaded from the database. However, this would typically be the same as
request.user, which is usually already loaded.

Now, request.user is used when it has the same id as registration.user
would have, saving one query. When this would result in the wrong user,
registration.user is used as before.