RegistrationStepMixin: Make event a cached_property
This makes sure that it is lazily loaded only when needed. In
particular, when handling the final check form, it is not needed, so
this saves one model load (note that the RegistrationStatusService does
load the event, but since it needs to lock the event, this would load
the event twice previously).