models: Disable QuerySet.update everywhere
When using QuerySet.update, the updated_at field is not automatically
updated (this is not implemented in Django, partly because of backward
compatibility). Instead of silently not changing the updated timestamp,
just raise an error when update is used for now.

This is done by adding a mixin to each model's QuerySet class. For some
models, this meant adding a custom QuerySet and/or Manager class first,
which are empty (since the needed behaviour comes from the mixin). This
also makes it easy to actually implement the timestamp updating when we
need it later.