poetry: Add mysqlclient as optional dependency
This makes sure that *if* it is manually installed, it is not removed by
`poetry install --no-dev`. This is not currently relevant, since
mysqlclient and its dependencies are not listed as dev packages, so
poetry will already leave them alone, but making this explicit is
probably better.

It might be better to add this as an "extras" dependency, so it can be
installed by `poetry install --extras`, but it seems that that does not
currently work:

https://github.com/python-poetry/poetry/issues/1145