EventAdmin: Order by ascending start_date
The Event model default is descending, which make sense if you have a
lot of events, but with just two events, ascending is more useful.

In the future, the Django admin should probably be configured to only
show future events by default, so then ascending still makes sense
(probably both for the default future only view, as well as for an
explicit show past events view).

This leaves the Event model default ordering at descending, which is
used in e.g. filter interfaces etc. that cannot themselves be easily
filtered.