RegistrationStatusService: Fix minor race condition
This condition could occur when the *same* registration is finalized
twice. Because the registration status was checkd before taking the
lock, if the second request happened between the check and taking the
lock, both requests would be handled. If the first one made the event or
one of the options full, that could cause the second one to move the
event to the waitinglist.