admin: Allow sorting on derived fields too
This adds some additional annotations to models and admin classes that
tell the admin how to sort on derived fields (like user's full name, or
registration timestamp with milliseconds displayed, etc.).

This does introduce some duplication between the actual value
calculation (Python expression) and the sort value (database
expression), and due to limitations also duplicates the sort value
across relations in two cases. Maybe this can be improved in Django at
some point.

Some of these derived fields are properties rather than methods. It
seems Django 2.2 cannot sort these yet (it does not see the
admin_order_field attribute on them), but this seems to be fixed in 3.0
or 3.1, so these attributes are added already.