ConsentLog: No cascaded delete when Registration is deleted
Deletion of ConsentLog entries is completely disallowed, so this cascade
relation would also prevent deleting registrations. Instead, just set
the registration field to NULL when the related registration is deleted,
to keep the registration of consent. The consent description still notes
the name of the event consent was given for, for manual auditing.

Note that deleting a user is still disallowed, since setting that to
NULL really invalidates the consent, so this should probably be handled
differently.