registrations: Modify models to support pending registrations
These are registrations that are not admitted immediately when
registration is finalized, but kept as pending to be admitted by some
later date using some external mechanism (e.g. a lottery or selection
process).

This introduces a new PENDING state, admit_immediately fields on Event
and RegistrationFieldOption that control whether a registration should
be admitted immediately or not and a pending_mail_text on Event.

This commit just makes the model changes, the behavioral changes will be
made in a separate commit.