registrations: Make final check form tests more specific
Previously, these would check for the presence of a form tag to see
whether registration could be finalized. Now, this checks for the
presence of an agree input and submit button, which is a stricter check,
and prepares for refactoring the templates in a way where a (possibly
empty) form tag is always generated.