RegistrationStart: Make subclass of RegistrationStepMixin
This also includes the start step in REGISTRATION_STEPS and lets it use
step_base.html.

There is a small change in behavior when doing a POST to the start view
when the registration is already in PREPARATION_COMPLETED, which now
redirects to finalcheck immediately rather than to the first options
step. This affects one testcase, which is updated accordingly.