RegistrationConfirmation: Make subclass of RegistrationStepMixin
This also includes the start step in REGISTRATION_STEPS and lets it use
step_base.html.

Because this is now the last step, and it has no form that needs a
success_url, the special handling for success_url that was in place for
FinalCheck is no longer needed and thus removed.