RegistrationOptionsForm: HTML-escape field titles
These would would be passed as labels to the generated fields, which
would render them as HTML directly. However, in other places (i.e.
rendered through templates) any HTML in these titles would be escaped,
making things inconsistent. Now, HTML is escaped everywhere, which is
the easiest and probably sufficient.

Note that field's help_texts are not escaped, since those are only
rendered as helptexts, and using HTML in there can be useful.