RegistrationOptionsForm: Allow omitting fields with unmet dependencies
Previously, fields with unmet dependencies would not be saved to the
database, but a value would still need to be specified for them in the
POST data. In practice, only CHOICE-typed fields would be used with
depends, which currently always send a value (there is no browser-side
javascript yet to hide them, and there is no "None" value currently).
However, e.g. empty STRING-typed values would prevent validating the
form, even when they had unmet dependencies.

This fixes this by removing the validation errors if a field turns out
to have unmet dependencies. Note that we cannot remove the required
attribute before calling the superclass clean (which generates the
errors in the first place), since we need access to the cleaned data for
the depended-on options to decide if a dependency is met or not.

This also adds some tests.

Note that this is only the server-side handling, currently the browser
will still complain about these fields because they have required set in
the HTML. This must be fixed later in javascript (see #100).