RegistrationField: Allow making fields optional
This adds a required attribute, which defaults to True as before.

For checkboxes, this changes what required=True means: a required
checkbox must now be checked to complete the registration, like the
Django form field.

In addition, this adds additional checking on field values when going to
PREPARATION_COMPLETED. Before, the presence of a RegistrationFieldValue
was sufficient, now the value itself is also checked.