RegistrationField: Fix value of UNCHECKBOX field type
In the first draft, the field type was called CHECKEDBOX, but this was
changed to UNCHECKBOX before committing. However, the database value was
still left at the old value when commiting. This makes that consistent.

Note that there is no data migration added, since the old value is not
actually in use anywhere.