templates/registered_event: Also link to event website for cancellation
Our upcoming event, 1672, has event-specific exceptions on the
cancellation policy, so it is somewhat misleading to (only) point at the
house rules. Replacing the link just for this event is too hardcoded, so
this is a compromise that at least shows the link to the house rules and
the event website.