Event: Add user-specific is_full annotation
This predicts whether the option is full for a particular user, by
looking at the options that are available to that user (rather than just
the total number of slots of the event, which is what the full property
indicates).

This takes into account fullness of options and invite_only of options,
but leaves out dependencies.