RegistrationOptionsForm: Fix display of previously saved unchecked checkboxes
When a checkbox was previously saved with an unchecked value ('0'), it
would be displayed as checked when displaying the options form. It was
assumed that Django would convert the string value ('0' or '1') to the
appropriate value, but it turns out this only applies to values
submitted through the form, not to the initial values, which must be
True/False already.

This adds the appropriate conversion.