RegistrationOptionsForm: Allow clearing an optional image field
This fixes two problems. Firstly, the required option for form fields
was set after creating the fields, but an ImageField only forwards its
required attribute to its widget in the constructor. This means
validation would work ok, but the widget would not know the field was
optional, and not show the "clear" checkbox. This is fixed by passing
the required value (and for consistency, also label and help_text) to
the constructor instead of setting it later.

Secondly, when the clear checkbox was checked, the value returned by the
form was False, while the model field would need the empty string to
clear (otherwise it would end up storing a filename of "False"). This is
fixed by converting the value on save.