templates: Improve account settings layout and styling
This introduces a left bar with a menu, similar to the progress menu of
the registration procedure, also highlighting the currently selected
item.

This also moves some of the duplicate form-generating code to the base
template (again similar to the registration steps), simplifying the
other templates (except for the email addresses template, which needs to
show multiple forms).