EventAdmin: Allow exporting active registrations for an event
This applies quite a bit of customization to django-import-export
resource to get the event options to show up in the export, and
duplicates some code from django-import-export to make a custom admin
action to export registrations for a single event. This might not be the
perfect approach, but at least it works for now, and going through a
django-import-export Resource class probably makes it easier to use the
same export in other places too.