ConsentLog: Add and use log_consent helper method
This unifies all places that create ConsentLog objects to do so in the
same way, reducing some duplicate code.

For the registration and email preferences forms, this now also includes
the form label in the consent description, like was already done for
medical consent. This was not originally so important, but with recent
changes to the labels and help texts for these fields, the label should
really be logged too.

To prevent directly creating ConsentLog objects, this also overrides the
`create()` method on the manager.