payments: Add app
This adds models to track payments (both manual and automatic using
Mollie), a view and service to handle webhook status updates from
Mollie, a service to initiate payments with Mollie and tests for all
these.

None of these are already used, this will be commited later.

This implements part of #21.