registrations: Implement online payment flow
This adds a PaymentStatus view that can be reached from the dashboard,
which shows the current payment status of a registration and allows
paying it if not (completely) paid yet. This also adds a PaymentDone
view to which the user is redirected after completing (or failing) the
payment.

This does not integrate the payment in the registration procedure yet
(which would make sense for admit_immediately registrations), it is
currently only a separate action after registration.

This also still lacks unittests.

This fixes #21.