RegistrationConfirmation: Update payment info
This updates both confirmation page and the email, removing the bank
account info, since payments can now be done through mollie (including
bank transfers). The payment url is also added to the confirmation.
Finally, a remark about no longer requiring waitinglist payments as an
exception is removed, as the online payment now no longer allows payment
for waitinglist registrations (though this still needs a change in the
house rules if we want to really do it like this).