RegistrationAdmin: Add admin action for changing status
This makes it easier to do bulk updates to change pending registrations
to be registered, cancelled or on the waitinglist.

This is still quite rough, does not update full status properly, needs
to support more state changes, etc. but it works for this particular
case now, when used with care.