PaymentService: Reorder elements of payment description
It seems that using Mollie, the description as it appears on the bank
statement is limited. For example iDeal using Rabobank only shows 35
characters (with about as much being used for mollie identifiers). This
puts the registratinon number first, which should generally be short and
unique identifies the registration if needed, then the event name, which
is most informative for the user, and then the name, which is probaly
still useful even if truncated.