Registration: Give cancelled registrations a zero price
In practice, a part of the price might still be owed, but this should be
independently stored elsewhere later (this is part of #67), the original
price should no longer matter.

This also changes payment_status so that registrations with payments are
never considered free, since these can now be cancelled registrations
that still need to be refunded.