PaymentStatus: Strike through priced options for cancelled registrations
Since these no longer contribute to the price, it would be confusing to
show them as-is, but to still allow seeing what the original price was,
show them striked-through.

This does somewhat duplicate business logic in the template, but this
seems to be the most pragmatic way to achieve this.