Registration: Convert amount_due and payment_status to annotations
These were previously cached properties rather than annotations because
in an earlier attempt there was a bug in Django when comparing
Case-wrapped DecimalValues (amount_due) in the payment_status
annotation. That problem has now be diagnosed and a workaround is
applied, allowing the use of an annotation which will allow filtering on
payment status in queries.