RegistrationPriceCorrection: Add and use new model
This can be used to add manual price corrections to registrations. These
can be limited to active or cancelled registrations, to allow
corrections that are (like regular options) dropped when the
registration is cancelled, but also allow corrections that still/only
apply to a cancelled registration (i.e. a cancellation fee).