PaymentAdmin: Make it easier to provide just transaction date
The field also stores the time, which is mostly relevant for mollie
payments. For manual payments, this replaces the widget with a regular
date widget, so you only need to provide the transaction date (though
there is a datetime format set, so you *can* also provide a time if you
want).