LimitForeignKeyOptionsMixin: Add generalized LimitDependsMixin
This adds a new class to the (new) arta.common.admin package, which is
a generalization of the more specific LimitDependsMixin. The generalized
version can be used to limit the options for any ForeignKey field in the
admin, based on another field of the edited class, instead of just the
depends field based on the event.