RegistrationFieldValue(Inline)Admin: Improve editing of values
This limits the choices for some dropdowns based on other fields, and
limits editing of some values in inlines where they choices cannot be
properly limited (and adding a change link from the inline to the
regular admin where the option field is properly limited).

Additionally, makes some fields readonly after initial creation of
a RegistrationFieldValue. For the inline, this needed to create two
separate inlines, one for viewing the values with all fields readonly,
and one for adding new values. This trick was taken from
https://stackoverflow.com/a/28149575/740048

Note that another answer to that post has a simpler answer disabling
change permission on the inline to make all fields readonly, but that
also changes the "change" links into "view" links, which is not really
intuitive, so this more complex solution is used instead.