TestRegistrationForm: Refactor test_change_options
This stores the previous request in a variable, so the code explicitly
shows what is changed compare to the previous request, removing some
duplication (and preparing for adding more similar requests in the
future).

This also adds one test: that an image field keeps its value even when
no new image file is uploaded.