RegistrationOptionsForm: Allow changing active registrations
This decides which fields can be edited based on the allow_change_until
field property. Unchangeable fields are shown as read-only using
SpanWidget, and fields depending on unchangeable and non-selected
options are omitted completely (since these can never become active).

This also changes the dependency handling javascript to ignore
dependencies that are missing, since that means the dependency is
satisfied and unchangeable.

This implements part of #71 and #79.