registrations: Allow registration forms for active registrations
This essentially allows editing registrations by allowing the forms to
be used with active registrations. This is not the complete
implementation yet, since there is no link to start editing yet, and the
start and end UI need more tweaking.

This also modifies some tests to work with active registrations too. To
make this work, all fields have their allow_change_until set so they
can be modified.

This implements part of #71 and #79.