RegistrationStepMixin: Refactor current step handling
This wraps access to the REGISTRATION_STEPS constant into a new steps
property, which prepares for omitting steps in certain conditions in
a future commit. It also exposes a new current_step property to
immediately return current step info, instead of relying on other code
to look up this info based on the current step number.