registrations: Improve UI for changing registrations
This customizes the steps shown on the left when editing a registration
to omit the final check and use a different start and end view.

To decide whether the current view is about a new registration or
changing an existing one, it looks at a newly added `change' property in
`REGISTRATION_STEPS` or at the registration status for views that are
used in both. That same property is then used to decide which steps to
display.

This implements part of #71 and #79.