Registration: Add helper properties related to active options
This adds active_options to return a list of active options, which
replaces most uses of the options reverse relation (except notably
PaymentStatus, which should also show priced inactive options). This
property can be prefetched using prefetch_active_options().

This renames options_by_name to active_options_by_name and lets it use
active_options instead of options.

This adds active_options_by_section, which returns active_options but
grouped by section. This property is needed to allow easy access to this
from the registration_options template snippet. It would previously use
a queryset method for this, but since active_options returns a list
instead of a queryset (since prefetch to_attr produces a list), the
queryset method is no longer accessible to templates. To facilitate this
change, the group_by_section queryset method is moved to become
a group_by_section queryset method is moved to become a group_by_section
queryset method is moved to become a group_by_section queryset method is
moved to become a class method on RegistrationFieldValue as well.