RegistrationsTable: Improve layout
This adds horizontal lines and a bit or padding, making the table easier
to read.

Turning this into a bootstrap table would probably be even
nicer, but it seems quite hard to customize the table HTML output (which
is generated by django-import-export, which uses tablib which uses
PyMarkup2 and just hardcodes plain HTML).