RegistrationNotifyService: Fix options in confirmation mail
In commit b615f9a (Registration: Add helper properties related to active
options), a list of active options was passed to the template instead of
a queryset of all options, but that list did not have the
group_by_section method needed, so no options were shown.

This changes the view to pass an already-grouped list of options to the
template, so the options are displayed as intended again.