RegistrationStepMixin: Fix redirect to edit_start
This accidentally still used the registration id instead of the event
id. This was not caught in the tests, presumably because the primary
keys for the event and registration happen to be the same (when using
sqlite the autogenerated pk's are probably reset on each testcase). When
using Mysql, higher keys were used and the test failed.