RegistrationStatusService: Better handle invites and depends
Previously, invites and dependencies on options, and invites on fields
were not properly checked, meaning there was no validation error when
unavailable options were selected, and there *was* a validation error
when an invite-only field was missing.

This fixes these things by doing a much more thorough check, and also
updates the testcases to check all of these.