RegistrationsTableDownload: Make prices numeric in spreadsheet
Previously, these were text fields including a currency symbol,
preventing calculations with this fields. Now, the widget for these
fields is replaced by a DecimalWidget when exporting a spreadsheet, to
expose the original number to import_export/tablib. In addition, to work
around a bug (or missing feature) in the tablib ODS export, this
switches from ODS to XLSX export.

The exported cells are now plain numbers, not yet currency-formatted
cells, since it seems import_export and tablib do not support this at
all.