MonetaryResourceWidget: Add self argument to methods
These methods were lacking a self argument. However since this  widget
was never instantiated but passed as a class to the resource fields,
calls to these methods were implicitly done as static methods, so this
did not cause any problems.

This commit fixes both: Add a self argument and pass instantiated
objects.