RegistrationOptionField: Let UI handle option dependencies too
Until now, fields were hidden by javascript when their dependencies were
not fuilfilled, but dependencies on individual options were only checked
in the backend.

This modifies RegistrationOptionField and introduces a new widget to set
the data-depends-name and data-depends-value attributes on the generated
option HTML tag to express these dependencies. The existing javascript
turns out to be general enough to already handle everything needed with
this, so no changes to the javascript are needed.