RegistrationStatusService: Check full flags even when slots is None
Recently this case occurred in practice: an option which was limited,
but used a lottery so did not have slots defined. After the lottery,
"full" was checked in the expectation that this would put subsequent
registrations on the waiting list. This is also what the registration UI
predicted, but the actual registration would then *not* end up on the
waiting list.

This was because finalize_registration would only check the full flag
for options with slots. Now, all options are checked.