Event: Support invitee_group with its own open date
This adds an `invitee_group` attribute to an event, which lists users that
can have their own `invitee_registration_opens_at` attribute. This
allows opening up registrations early for some group, or completely
hiding an event for all but the invitees.

The business logic of this is implemented exclusively in the
`EventQueryeset.for_user()` annotations, by renaming the existing
`registration_opens_at` attribute to `public_registration_opens_at` and,
using annotations, select either `public_registration_opens_at` or
`invitee_registration_opens_at` (based on whether the user is member of
the invitee group) and annotate the result as `registration_opens_at`.
All other logic can just use this annotation like it did the old
property and works without changes.

The only place where the old attribute was used without the annotations
is in generating the caching key for the FinalCheck view, so this needed
one more change. And of course some test code needed to be updated.