Dashboard: Show correct opening date when timezone is not UTC
When the opening timestamp has a 00:00 time, only the date would be
shown. But this used the Python builtin `datetime.date()` method, which
returned the date in UTC (since the datetime object has a UTC date).
Now, this uses the "date" template filter, which instead uses the
configured local timezone instead, showing the right date.